Moved keychain code to separate python file
